[NayNay] File Restructure: Accounts Restructure
- refactoring file structure + flow of accounts in cli/tui
