+++ src/accounts/constants.ts
@@ -1,0 +1,12 @@
+export const IMPORT_CONTENT = {
+  seed: {
+    name: 'seed',
+    message: 'Enter seed:',
+    invalidSeed: 'Seed provided is not valid'
+  },
+  path: {
+    name: 'path',
+    message: 'derivation path:',
+    default: 'none',
+  }
+}
